support for bluetooth beacons
For now, beacons are recognized via their MAC address, because this seems
easier than relying on the dynamic list of IDs provided in the payload.

To support multi-beacons (those that support several beacon protocols)
with the same MAC address, the payload parser type is added to make the
beacons distinguishable.

also fix some inconsistencies when sending battery and offline messages

fixes #6