dynamic message configuration
Introduce a first iteration for sending dynamic messages. For now, the only
new functionality is the introduction of new message formats CSV, JSON and
YAML in addition to the existing PLAINTEXT format.

The new formats send only a single message per selected message items.
This means that if several conditions apply, their content is sent
together in the same message.