replace alarm receiver logic by WorkManager
While not all functionality of the old alarm based functionality can be
done with the WorkManager logic (especially scheduling with < 15 min
interval), the library offers full persistence over restarts and app
terminations by the OS, so the message scheduling should no longer stop
randomly.

Another benefit of using WorkManager is that the app no longer needs to
manually listen to network changes if it is currently not registered to
an allowed network, as the library will do that for us.

The issue with scheduling being limited to at least 15 minutes is
somewhat mitigated by scheduling the same work with different IDs and
different starting times, though only time will tell if those starting
time differences will actually be retained once the device has entered
and left doze mode.

Another drawback of this change is that the next run of the schedule
can no longer be properly calculated. Instead, during each execution,
the current interval is added to the current timestamp to estimate the
next execution. But as even before no guarantee was made for exact
execution times, this should be ok.

Side note: As the new schedule estimation had to be added to the
message contest anyway, I took the opportunity and separated the
calculation logic from the PoJos more clearly.