avoid race conditions in Scheduler
Previously if two subsequent runNow() calls were triggered in parallel
while the next planned schedule was still in the future, both workers
would be re-scheduled "now", and would at the end of their execution
notice that the next planned schedule was still in the future and
therefore not schedule another worker run. This caused the scheduling
to come to a halt.

To avoid this:
- synchronize all scheduling actions
- switch to next worker based on current worker
- check if the next worker is already done to detect if re-scheduling
  is required