keep preference listener to react to consent change
The change of the consent flag happens in a dialog, i.e. when the
main activity is paused. To still react to this and register the
network callback during initialization, we need to keep the
change listener registered. It should be fine to only unregister
on destroy: if the process is really killed, then the shared
preferences instance will also be gone, so it should not matter
that we did not call unregister on it.