hopefully really fix wifi detection issues #70
The main issue is that wifi manager sometimes returns a disconnected
WifiInfo object rather than null, which was wrongly detected as being
connected to a new "<unknown ssid>" network. This is addressed by
additionally checking the supplicant state of the network for state
COMPLETED.

Also for SDK < 31 we cannot really rely on the WifiManager giving
data in sync to the Callback events, as we have no guarantees about
that from AOSP. The best we can do is (similar to the Broadcast
solution for SDK < 23), regardless of event type, always to check
the current SSID and assume connect when it's present and
disconnect when it's absent.

Since this causes duplicate events, the last known SSID is statically
persisted in the WifiEventConsumer, and the event is only propagated
if the SSID actually changes.