check the wifi before publishing in Android 11 and older
Since the network callback solution does not work due to race conditions,
we additionally read the wifi during publishing again. Hopefully, this
should give us the best of both worlds.

Also, do not store the connected wifi into a preference, since we do not
want it to survive app restarts - who knows how long the app was not
running, most likely the wifi has changed in the meantime.

relates to #70